Improve the  script

+++ tool/check-agda.sh
@@ -6,20 +6,28 @@ script_dir=$(dirname "$script")
 root_dir=$(dirname "$script_dir")
 cd "$root_dir"
 
-# change into the agda base-lib dir (location of `.agda-lib` file)
-cd base/agda
-
-# check options
+# consume parameters
 while getopts "r" opt; do
   case $opt in
     r)
-      rm -r _build
+      rebuild=true
       ;;
     \?)
       exit 1
       ;;
   esac
 done
+
+# change into the directory containing the Agda modules
+ARG1=${@:$OPTIND:1}
+cd "$ARG1"
+echo "Building Agda files in $(pwd)"
+
+# Delete `_build` folder iff the user wants to rebuild everything
+if [ "$rebuild" = true ] ; then
+    echo "Removing _build directory"
+    rm -r _build
+fi
 
 # check all agda files
 for f in $(find . -name "*.agda")
